PD-2334: Refresh badge (#1383)
* darkmode

* changeset

* break out styles and types
