+++ .changeset/friendly-jokes-pull.md
@@ -1,0 +1,5 @@
+---
+'@leafygreen-ui/badge': major
+---
+
+Updates `Badge` for dark mode brand refresh.
