+++ packages/badge/package.json
@@ -20,8 +20,12 @@
     "access": "public"
   },
   "dependencies": {
-    "@leafygreen-ui/lib": "^9.3.0",
-    "@leafygreen-ui/palette": "^3.4.0"
+    "@leafygreen-ui/lib": "^9.4.1",
+    "@leafygreen-ui/palette": "^3.4.0",
+    "@leafygreen-ui/tokens": "1.3.1"
+  },
+  "peerDependencies": {
+    "@leafygreen-ui/leafygreen-provider": "^2.3.2"
   },
   "gitHead": "dd71a2d404218ccec2e657df9c0263dc1c15b9e0",
   "homepage": "https://github.com/mongodb/leafygreen-ui/tree/main/packages/badge",
