+++ packages/badge/src/Badge.story.tsx
@@ -1,10 +1,17 @@
 import { ComponentStory, Meta } from '@storybook/react';
 import React from 'react';
 import Badge from './Badge';
+import defaultArgTypes from '../../../stories/defaultArgTypes';
 
 export default {
   title: 'Components/Badge',
   component: Badge,
+  args: {
+    darkMode: false,
+  },
+  argTypes: {
+    darkMode: defaultArgTypes.darkMode,
+  },
 } as Meta<typeof Badge>;
 
 const Template: ComponentStory<typeof Badge> = args => <Badge {...args} />;
