+++ packages/badge/tsconfig.json
@@ -15,6 +15,12 @@
     },
     {
       "path": "../palette"
+    },
+    {
+      "path": "../tokens"
+    },
+    {
+      "path": "../leafygreen-provider"
     }
   ]
 }
